Add readme

+++ src/main.rs
@@ -1,9 +1,30 @@
+use clap::Parser;
 use futures::StreamExt;
 use indicatif::{MultiProgress, ProgressBar, ProgressStyle};
+use inquire::MultiSelect;
 use std::collections::HashMap;
-use std::error::Error;
+use std::fmt::{Display, Formatter};
 use std::path::PathBuf;
 use walkdir::WalkDir;
+
+#[derive(Parser)]
+#[clap(
+    version = "0.0.1",
+    about = "A tool to clean up node_modules folders in a project."
+)]
+pub struct Arguments {
+    #[clap(
+        short,
+        long,
+        default_value = "2",
+        help = "How deep to search for projects"
+    )]
+    pub depth: usize,
+    #[clap(short, long, help = "Silent mode", default_value = "false")]
+    pub silent: bool,
+    #[clap(short, long, help = "Interactive mode", default_value = "false")]
+    pub interactive: bool,
+}
 
 #[derive(Clone)]
 pub enum Status {
@@ -42,6 +63,12 @@ impl Project {
     }
 }
 
+impl Display for Project {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.path.display().to_string().split_at(2).1)
+    }
+}
+
 fn identify_projects(depth: usize) -> Vec<Project> {
     let mut projects = Vec::new();
     let mut walker = WalkDir::new(".").max_depth(depth).into_iter();
@@ -70,17 +97,13 @@ fn identify_projects(depth: usize) -> Vec<Project> {
 }
 
 async fn status_poller(rx: async_channel::Receiver<StatusUpdate>, projects: Vec<Project>) {
-    // Create a multi-progress to manage multiple lines of output
     let multi_progress = MultiProgress::new();
 
-    // Prepare to store our projects and associated progress bars
     let mut project_map: HashMap<PathBuf, Project> =
         projects.into_iter().map(|p| (p.path(), p)).collect();
 
-    // A map of path -> ProgressBar
     let mut bars = HashMap::new();
 
-    // Create a progress bar per project
     for project in project_map.values() {
         let pb = multi_progress.add(ProgressBar::new_spinner());
         pb.set_message(format!("Waiting: {}", project.path().display()));
@@ -90,20 +113,18 @@ async fn status_poller(rx: async_channel::Receiver<StatusUpdate>, projects: Vec<
                 .template("{spinner} {wide_msg}")
                 .unwrap(),
         );
-        // Spin at regular intervals
+
         pb.enable_steady_tick(std::time::Duration::from_millis(100));
 
         bars.insert(project.path(), pb);
     }
 
-    // Keep receiving status updates
     while let Ok(update) = rx.recv().await {
         if let Some(project) = project_map.get_mut(&update.path) {
             project.update_status(update.status.clone());
             if let Some(pb) = bars.get(&update.path) {
                 match update.status {
                     Status::Waiting => {
-                        // Typically won't happen once we start, but let's handle it anyway
                         pb.set_message(format!("{}", project.path().display()));
                     }
                     Status::Deleting => {
@@ -127,7 +148,7 @@ async fn delete_projects(projects: Vec<Project>, tx: async_channel::Sender<Statu
             .into_iter()
             .map(|p| delete_project(p.path(), tx.clone())),
     )
-    .buffer_unordered(4)
+    .buffer_unordered(8)
     .collect::<Vec<_>>()
     .await;
 }
@@ -164,11 +185,32 @@ async fn delete_project(path: PathBuf, tx: async_channel::Sender<StatusUpdate>) 
 
 #[tokio::main]
 async fn main() {
-    let projects = identify_projects(2);
+    let args = Arguments::parse();
+
+    let mut projects = identify_projects(args.depth);
+
+    if projects.is_empty() {
+        if !args.silent {
+            println!("No node_modules folders found.");
+        }
+        return;
+    }
 
     let (tx, rx) = async_channel::unbounded();
 
-    tokio::spawn(delete_projects(projects.clone(), tx));
+    if args.interactive {
+        match MultiSelect::new("Select projects to clean.", projects.clone()).prompt() {
+            Ok(ans) => projects = ans,
+            Err(_) => {
+                return;
+            }
+        }
+    }
 
-    status_poller(rx, projects).await;
+    if !args.silent {
+        tokio::spawn(delete_projects(projects.clone(), tx));
+        status_poller(rx, projects).await;
+    } else {
+        delete_projects(projects, tx).await;
+    }
 }
